pyproject.toml
[project]
name = "extractors-uv"
version = "0.1.0"
description = ""
readme = "README.md"
authors = [
  {name = "DevScanr Team", email = "info@devscanr.com"}
]
requires-python = ">=3.12,<3.13"
dependencies = [
  "beautifulsoup4>=4.13.3",
  "emoji>=2.14.1",
  "lingua-language-detector>=2.0.2",
  "markdown>=3.7",
  "spacy>=3.8.4",
  "urlextract>=1.9.0",
]

[dependency-groups]
dev = [
    "mypy>=1.15.0",
    "psutil>=7.0.0",
    "pytest>=8.3.5",
    "ruff>=0.9.9",
    "types-beautifulsoup4>=4.12.0.20250204",
    "types-markdown>=3.7.0.20241204",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.poe.tasks]
build = "uv build --wheel"
tc = "uv run mypy ."
test = "uv run pytest --disable-warnings"
lint = "uv run ruff check"

[tool.mypy]
strict = true
disallow_untyped_calls = false
no_implicit_reexport = false
exclude = [
  "^ignore",
]

[tool.ruff.lint]
select = ["F", "B", "UP", "RUF"]
ignore = ["E701", "E711", "E712", "E731", "E741", "F541", "RUF001", "RUF002", "RUF003", "RUF021"]
per-file-ignores = {"__init__.py" = ["F403"]}

[tool.pytest.ini_options]
testpaths = ["extractors"]
# addopts = "--ignore=ignore"
python_files = "*_test.py"
#python_classes = ["*__*"]
#python_functions = ["*__*"]
